[poke-73] reorganice e2e in different files
